refactor(settings): unify integrations panel
1. Panel consolidation
   - Merge CGP + IR into IntegrationsPanel
   - Remove IndentRainbowPanel (git mv)
   - Wire new panel in Configurable

2. Swatch redesign
   - Single-row adaptive-width layout
   - Checkmark selection indicator
   - Contrast-aware check color

3. Cleanup
   - Remove CGP code from ElementsPanel
   - Update kover exclusion class name

Impact: single Integrations group for all
third-party plugin toggles; swatches fill
available width automatically
