+++ build.gradle.kts
@@ -137,8 +137,8 @@ kover {
                     "dev.ayuislands.accent.CachedMacReader*",
                     "dev.ayuislands.accent.SystemAccentProvider*",
                     "dev.ayuislands.accent.SystemAppearanceProvider*",
-                    // Pure-rendering UI panel (Indent Rainbow settings)
-                    "dev.ayuislands.settings.IndentRainbowPanel*",
+                    // Pure-rendering UI panel (Integrations settings)
+                    "dev.ayuislands.settings.IntegrationsPanel*",
                     // IDE glue (thin event listeners, startup activity)
                     "dev.ayuislands.AyuIslandsStartupActivity*",
                     "dev.ayuislands.AyuIslandsLafListener*",
