+++ src/main/kotlin/dev/ayuislands/settings/ColorSwatchPanel.kt
@@ -2,7 +2,6 @@ package dev.ayuislands.settings
 
 import com.intellij.ui.ColorUtil
 import dev.ayuislands.accent.AccentColor
-import java.awt.AlphaComposite
 import java.awt.BasicStroke
 import java.awt.Color
 import java.awt.Cursor
@@ -16,13 +15,12 @@ import java.awt.event.MouseEvent
 import java.awt.geom.RoundRectangle2D
 import javax.swing.JComponent
 import javax.swing.JPanel
-import javax.swing.UIManager
 
-/** A 2x6 grid of rounded color swatches with a checkmark on the selected color. */
+/** A single-row grid of rounded color swatches with checkmark on the selected color. */
 class ColorSwatchPanel(
     colors: List<AccentColor>,
     private val onColorSelected: (AccentColor) -> Unit,
-) : JPanel(GridLayout(GRID_ROWS, GRID_COLS, GRID_GAP, GRID_GAP)) {
+) : JPanel(GridLayout(1, colors.size, GRID_GAP, 0)) {
     var selectedColor: String = ""
         set(value) {
             field = value
@@ -70,6 +68,9 @@ class ColorSwatchPanel(
             try {
                 g2.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON)
 
+                val color = Color.decode(accent.hex)
+                val isSelected = accent.hex.equals(selectedColor, ignoreCase = true)
+
                 val shape =
                     RoundRectangle2D.Float(
                         0f,
@@ -80,17 +81,21 @@ class ColorSwatchPanel(
                         ARC_RADIUS,
                     )
 
-                val color = Color.decode(accent.hex)
                 g2.color = color
                 g2.fill(shape)
 
                 if (!isEnabled) {
-                    g2.composite = AlphaComposite.getInstance(AlphaComposite.SRC_OVER, DISABLED_OVERLAY_ALPHA)
-                    g2.color = UIManager.getColor("Panel.background") ?: background
+                    g2.composite =
+                        java.awt.AlphaComposite.getInstance(
+                            java.awt.AlphaComposite.SRC_OVER,
+                            DISABLED_OVERLAY_ALPHA,
+                        )
+                    g2.color = javax.swing.UIManager.getColor("Panel.background") ?: background
                     g2.fill(shape)
+                    g2.composite = java.awt.AlphaComposite.getInstance(java.awt.AlphaComposite.SRC_OVER, 1f)
                 }
 
-                if (accent.hex.equals(selectedColor, ignoreCase = true)) {
+                if (isSelected) {
                     drawCheckmark(g2, color)
                 }
             } finally {
@@ -105,25 +110,22 @@ class ColorSwatchPanel(
             val checkColor =
                 if (ColorUtil.isDark(background)) Color.WHITE else Color(DARK_TEXT_R, DARK_TEXT_G, DARK_TEXT_B)
             g2.color = checkColor
-            g2.stroke = BasicStroke(STROKE_WIDTH, BasicStroke.CAP_ROUND, BasicStroke.JOIN_ROUND)
+            g2.stroke = BasicStroke(CHECKMARK_STROKE, BasicStroke.CAP_ROUND, BasicStroke.JOIN_ROUND)
 
             val centerX = width / 2
             val centerY = height / 2
             val size = CHECKMARK_SIZE
 
-            // Check path: short left stroke then longer right stroke
             g2.drawLine(centerX - size, centerY, centerX - 1, centerY + size)
             g2.drawLine(centerX - 1, centerY + size, centerX + size, centerY - size + 1)
         }
     }
 
     companion object {
-        private const val GRID_ROWS = 2
-        private const val GRID_COLS = 6
         private const val GRID_GAP = 6
         private const val SWATCH_SIZE = 28
         private const val ARC_RADIUS = 6f
-        private const val STROKE_WIDTH = 2f
+        private const val CHECKMARK_STROKE = 2f
         private const val CHECKMARK_SIZE = 4
         private const val DARK_TEXT_R = 0x1F
         private const val DARK_TEXT_G = 0x24
